refactor[file]: move service mockdata
